refactor: isolate Visual Studio toolchain cache
