+++ tests/native/assert_cpp_layout.ps1
@@ -231,6 +231,7 @@ $msvcLeafFiles = [ordered]@{
         'PortableToolchainDiscovery.cpp', 'PortableToolchainDiscovery.hpp',
         'ToolchainDiscoveryPrimitives.cpp', 'ToolchainDiscoveryPrimitives.hpp',
         'VisualStudioEnvironment.cpp', 'VisualStudioEnvironment.hpp',
+        'VisualStudioToolchainCache.cpp', 'VisualStudioToolchainCache.hpp',
         'VisualStudioToolchainDiscovery.cpp', 'VisualStudioToolchainDiscovery.hpp'
     )
 }
